data/binance_data.py
# From provided, enhance fetch_klines for date ranges
from binance.client import Client
from datetime import datetime
import pandas as pd

class BinanceData:
    def __init__(self, api_key, api_secret):
        self.client = Client(api_key, api_secret)
    
    def get_historical_data(self, symbol, interval='1d', start_str=None, end_str=None, limit=1000):
        # Docs: Use startTime/endTime in ms for ranges
        start_time = int(datetime.strptime(start_str, '%Y-%m-%d').timestamp() * 1000) if start_str else None
        end_time = int(datetime.strptime(end_str, '%Y-%m-%d').timestamp() * 1000) if end_str else None
        klines = self.client.get_historical_klines(symbol, interval, start_str=start_str, end_str=end_str, limit=limit)
        df = pd.DataFrame(klines, columns=[...])  # As in provided
        return df[['timestamp', 'open', 'high', 'low', 'close', 'volume']]