requirements.txt
streamlit>=1.28.0
numpy>=1.21.0
pandas>=1.3.0
plotly>=5.0.0
scipy>=1.7.0
matplotlib>=3.5.0
alpaca-py>=0.43.2
python-binance>=1.0.32
ta-lib  # conda install -c conda-forge ta-lib
empyrical>=0.5.5  # For analyzers (Sharpe/SQN alternative to pyfolio)
